[Sprint 17][main-svc-events] Add countByCategoryId + move controller
