+++ main-service/src/main/java/ru/practicum/ewm/events/controller/EventsController.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.events;
+package ru.practicum.ewm.events.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
