+++ main-service/src/main/java/ru/practicum/ewm/events/storage/EventsRepository.java
@@ -8,4 +8,6 @@ import ru.practicum.ewm.events.model.Event;
 
 public interface EventsRepository extends JpaRepository<Event, Long>, QuerydslPredicateExecutor<Event> {
     Page<Event> findAllByInitiatorIdIs(Long userId, Pageable pageable);
+
+    long countByCategoryId(Long categoryId);
 }
